change pgjson for new schema
