+++ src/fluentd/code/out_pgjson.rb
@@ -52,10 +52,6 @@ module Fluent::Plugin
 
     def initialize
       super
-      # if @conn == nil: The connection has not been established yet.
-      # if @conn != nil: The connection has been established, but it could be a broken connection.
-      # @conn = nil
-      # @thread_lock = Mutex.new
       @last_reset_ts = 0
     end
 
@@ -74,66 +70,48 @@ module Fluent::Plugin
     end
 
     def init_connection
-      # This function is used to init the connection (first connecting).
-      # If @conn == nil, try to establish a connection, and set @conn. 
-      # If @conn != nil, skip this function.
-      # @conn could be a nil or non-nil value after this call.
+      # This function is used to create a connection.
+      thread = Thread.current
       begin
         log.debug "[pgjson] [init_connection] Connecting to PostgreSQL server #{@host}:#{@port}, database #{@database}..."
-        conn = PG::Connection.new(dbname: @database, host: @host, port: @port, sslmode: @sslmode, user: @user, password: @password)
+        thread[:conn] = PG::Connection.new(dbname: @database, host: @host, port: @port, sslmode: @sslmode, user: @user, password: @password)
       rescue PG::Error
         log.debug "[pgjson] [init_connection] Failed to initialize a connection."
-        if ! conn.nil?
-          conn.close()
-          conn = nil
+        if ! thread[:conn].nil?
+          thread[:conn].close()
+          thread[:conn] = nil
         end
       rescue => err
         log.debug "#{err}"
       end
-      conn
+    end
+
+    def reset_connection	
+      # This function try to fix the broken connection to database.	
+      # if conn == nil, call init_connection	
+      # if conn != nil, call conn.reset
+      thread = Thread.current
+      begin	
+        if timestamp - @last_reset_ts > @reset_connection_interval	
+          if thread[:conn].nil?	
+            log.debug "[pgjson] [reset_connection] Call init_connection."	
+            init_connection	
+          else	
+            log.debug "[pgjson] [reset_connection] Reset Connection."	
+            thread[:conn].reset	
+          end	
+        else	
+          log.debug "[pgjson] [reset_connection] Skip reset."	
+        end	
+      rescue => err	
+        log.debug "[pgjson] [reset_connection] #{err.class}, #{err.message}"	
+      ensure	
+        @last_reset_ts = timestamp	
+      end
     end
 
     def timestamp
        Time.now.getutc.to_i
-     end
-
-    def reset_connection
-      # This function try to fix the broken connection to database.
-      # if @conn == nil, call init_connection
-      # if @conn != nil, call @conn.reset
-      # This function must be protected by thread_lock to ensure thread safety.
-      begin
-        if timestamp - @last_reset_ts > @reset_connection_interval
-          if @conn.nil?
-            log.debug "[pgjson] [reset_connection] Call init_connection."
-            init_connection
-          else
-            log.debug "[pgjson] [reset_connection] Reset Connection."
-            @conn.reset
-          end
-        else
-          log.debug "[pgjson] [reset_connection] Skip reset."
-        end
-      rescue => err
-        log.debug "[pgjson] [reset_connection] #{err.class}, #{err.message}"
-      ensure
-        @last_reset_ts = timestamp
-      end
-    end
-
-    def shutdown
-      # begin
-      #   @thread_lock.lock()
-      #   if ! @conn.nil?
-      #     @conn.close()
-      #     @conn = nil
-      #   end
-      # rescue => err
-      #   log.debug "[pgjson] [shutdown] #{err.class}, #{err.message}"
-      # ensure
-      #   @thread_lock.unlock()
-      # end
-      super
     end
 
     def formatted_to_msgpack_binary
@@ -150,31 +128,32 @@ module Fluent::Plugin
 
     def write(chunk)
       log.debug "[pgjson] in write, chunk id #{dump_unique_id_hex chunk.unique_id}"
-      # @thread_lock.lock()
-      conn = init_connection
-      if ! conn.nil?
+      thread = Thread.current
+      if ! thread.key?(:conn)
+        init_connection
+      end
+      if ! thread[:conn].nil?
         begin
-          conn.sync_exec("COPY #{@table} (#{@tag_col}, #{@time_col}, #{@record_col}) FROM STDIN WITH DELIMITER E'\\x01'")
+          thread[:conn].exec("COPY #{@table} (#{@tag_col}, #{@time_col}, #{@name_col}, #{@index_col}, #{@record_col}) FROM STDIN WITH DELIMITER E'\\x01'")
           tag = chunk.metadata.tag
           chunk.msgpack_each do |time, record|
-            conn.put_copy_data "#{tag}\x01#{time}\x01#{record_value(record, conn)}\n"
+            thread[:conn].put_copy_data "#{tag}\x01#{time}\x01#{record.name}\x01#{record.index}\x01#{record_value(record)}\n"
           end
         rescue PG::ConnectionBad, PG::UnableToSend => err
           # connection error
-          conn = init_connection # try to reset broken connection, and wait for next retry
+          reset_connection # try to reset broken connection, and wait for next retry
           log.debug "%s while copy data: %s" % [ err.class.name, err.message ]
           retry
         rescue PG::Error => err
           log.debug "[pgjson] [write] Error while writing, error is #{err.class}"
           errmsg = "%s while copy data: %s" % [ err.class.name, err.message ]
-          conn.put_copy_end( errmsg )
-          conn.get_result
+          thread[:conn].put_copy_end( errmsg )
+          thread[:conn].get_result
           raise errmsg
         else
-          conn.put_copy_end
-          res = conn.get_result
+          thread[:conn].put_copy_end
+          res = thread[:conn].get_result
           raise res.result_error_message if res.result_status != PG::PGRES_COMMAND_OK
-          conn.close()
           log.debug "[pgjson] write successfully, chunk id #{dump_unique_id_hex chunk.unique_id}"
         end
       else
@@ -182,9 +161,10 @@ module Fluent::Plugin
       end
     end
 
-    def record_value(record, conn)
+    def record_value(record)
+      thread = Thread.current
       if @msgpack
-        "\\#{conn.escape_bytea(record.to_msgpack)}"
+        "\\#{thread[:conn].escape_bytea(record.to_msgpack)}"
       else
         json = @encoder.dump(record)
         json.gsub!(/\\/){ '\\\\' }
