+++ src/fluentd/run.sh
@@ -1,5 +1,4 @@
 for VAR in $(seq 1 10)
 do
-	curl -i -X POST -d "json={\"index\":$VAR, \"name\":\"piggy\"}" http://localhost:8888/testpg.cycle
+	curl -i -X POST -d "json={\"index\":$VAR,  \"metadata\":{\"name\":\"piggy\"}}" http://localhost:8888/testpg.cycle
 done
-
