+++ src/fluentd/src/fluent-plugin-pgjson/example.conf
@@ -1,26 +1,17 @@
 <source>
-  @type http
-  port 8888
-  bind 0.0.0.0
+  type forward
 </source>
 
-<match test.cycle>
-  @type stdout
+<match **>
+  type pgjson
+  host localhost
+  port 5432
+  sslmode require
+  database fluentd
+  table fluentd
+  user postgres
+  password postgres
+  time_col time
+  tag_col tag
+  record_col record
 </match>
-# <source>
-#   type forward
-# </source>
-
-# <match **>
-#   type pgjson
-#   host localhost
-#   port 5432
-#   sslmode require
-#   database fluentd
-#   table fluentd
-#   user postgres
-#   password postgres
-#   time_col time
-#   tag_col tag
-#   record_col record
-# </match>
